requirements.txt
--only-binary :all:
Flask>=2.0.1
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
joblib>=1.0.1
requests>=2.25.1
python-dotenv>=0.19.0
kaggle>=1.7.4.5
